requirements-base.txt
absl-py==0.11.0
aiofiles==0.6.0
alembic==1.4.3
aliyun-python-sdk-core==2.13.30
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-kms==2.14.0
arrow==0.17.0
astunparse==1.6.3
bcrypt==3.2.0
blis==0.7.4
cachetools==4.2.0
catalogue==1.0.0
certifi==2020.12.5
cffi==1.14.4
click==7.1.2
cloudpickle==1.6.0
crcmod==1.7
cryptography==3.2.1
cymem==2.0.5
decorator==4.4.2
defusedxml==0.6.0
ecdsa==0.14.1
fastapi==0.63.0
fastapi-permissions==0.2.7
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.32.0
gym==0.18.0
h11==0.11.0
h5py==2.10.0
httpcore==0.12.2
httplib2==0.18.1
httpx==0.16.1
idna==2.10
importlib-metadata==3.3.0
intervaltree==3.1.0
Jinja2==2.11.2
jira==2.0.0
jmespath==0.10.0
joblib==1.0.0
kafka-python==2.0.2
Keras-Preprocessing==1.1.2
korean-lunar-calendar==0.2.1
libcst==0.3.16
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
murmurhash==1.0.5
mypy-extensions==0.4.3
numpy==1.19.4
oauth2client==4.1.3
oauthlib==3.1.0
opt-einsum==3.3.0
ortools==9.0.9048
oss2==2.13.1
pandas==1.1.5
patsy==0.5.1
pbr==5.5.1
Pillow==7.2.0
plac==1.1.3
preshed==3.0.5
proto-plus==1.13.0
protobuf==3.15.8
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.9
pydantic==1.8.1
pyglet==1.5.0
PyJWT==1.7.1
PyMeeus==0.3.13
pyparsing==2.4.7
pypd==1.1.0
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.2.0
python-multipart==0.0.5
pytz==2020.4
PyYAML==5.3.1
redis==3.5.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rsa==4.6
schedule==0.6.0
scipy==1.5.4
sentry-asgi==0.2.0
sentry-sdk==0.19.5
sh==1.14.1
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.3.0
spacy==2.3.5
SQLAlchemy==1.3.22
sqlalchemy-filters==0.12.0
SQLAlchemy-Searchable==1.2.0
SQLAlchemy-Utils==0.36.8
srsly==1.0.5
starlette==0.13.6
statsmodels==0.12.1
tabulate==0.8.7
tenacity==6.3.1
termcolor==1.1.0
thinc==7.4.5 
typing-extensions==3.7.4.3
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.26.2
uvicorn==0.12.3
validators==0.18.2
wasabi==0.8.0


chardet==3.0.4
opencv-python==4.4.0.46
holidays==0.10.2
ray[rllib]==1.2.0
# websockets==8.1